Struct sysfs_gpio::PinPoller

pub struct PinPoller {
    // some fields omitted
}

Methods

impl PinPoller

fn get_pin(&self) -> Pin

Get the pin associated with this PinPoller

Note that this will be a new Pin object with the proper pin number.

fn new(pin_num: u64) -> Result<PinPoller>

Create a new PinPoller for the provided pin number

fn poll(&mut self, timeout_ms: isize) -> Result<Option<u8>>

Block until an interrupt occurs

This call will block until an interrupt occurs. The types of interrupts which may result in this call returning may be configured by calling set_edge() prior to makeing this call. This call makes use of epoll under the covers. If it is desirable to poll on multiple GPIOs or other event source, you will need to implement that logic yourself.

This function will return Some(value) of the pin if a change is detected or None if a timeout occurs. Note that the value provided is the value of the pin as soon as we get to hanlding the interrupt in userspace. Each time this fuction returns with a value, a change has occurred, but you could end up reading the same value multiple times as the value has changed back between when the interrupt ocurred and the current time.