Struct sysfs_gpio::Pin

pub struct Pin {
    // some fields omitted
}

Methods

impl Pin

fn new(pin_num: u64) -> Pin

Create a new Pin with the provided pin_num

This function does not export the provided pin_num.

fn with_exported<F: FnOnce() -> Result<()>>(&self, closure: F) -> Result<()>

Run a closure with the GPIO exported

Prior to the provided closure being executed, the Gpio will be eported. After the closure execution is complete, the Gpio will be unexported.

Example

gpio = Pin::new(24);
let res = gpio::with_exported(|| {
    println!("At this point, the Pin is exported");
    try!(gpio.set_direction(Direction::Low));
    try!(gpio.set_value(1));
    // ...
};

fn export(&self) -> Result<()>

Export the GPIO

This is equivalent to echo N > /sys/class/gpio/export with the exception that the case where the GPIO is already exported is not an error.

Errors

The main cases in which this function will fail and return an error are the following: 1. The system does not support the GPIO sysfs interface 2. The requested GPIO is out of range and cannot be exported 3. The requested GPIO is in use by the kernel and cannot be exported by use in userspace

Example

use sysfs_gpio::Pin;

gpio = Pin::new(24);
match gpio.export() {
    Ok(()) => println!("Gpio {} exported!", gpio.pin),
    Err(err) => println!("Gpio {} could not be exported: {}", gpio.pin, err),
}

fn unexport(&self) -> Result<()>

Unexport the GPIO

This function will unexport the provided by from syfs if it is currently exported. If the pin is not currently exported, it will return without error. That is, whenever this function returns Ok, the GPIO is not exported.

fn set_direction(&self, dir: Direction) -> Result<()>

Set this GPIO as either an input or an output

The basic values allowed here are Direction::In and Direction::Out which set the Pin as either an input or output respectively. In addition to those, two additional settings of Direction::High and Direction::Low. These both set the Pin as an output but do so with an initial value of high or low respectively. This allows for glitch-free operation.

Note that this entry may not exist if the kernel does not support changing the direction of a pin in userspace. If this is the case, you will get an error.

fn get_value(&self) -> Result<u8>

Get the value of the Pin (0 or 1)

If successful, 1 will be returned if the pin is high and 0 will be returned if the pin is low (this may or may not match the signal level of the actual signal depending on the GPIO "active_low" entry).

fn set_value(&self, value: u8) -> Result<()>

Set the value of the Pin

This will set the value of the pin either high or low. A 0 value will set the pin low and any other value will set the pin high (1 is typical).

fn set_edge(&self, edge: Edge) -> Result<()>

Set the edge on which this GPIO will trigger when polled

The configured edge determines what changes to the Pin will result in poll() returning. This call will return an Error if the pin does not allow interrupts.

fn get_edge(&self) -> Result<Edge>

Get the currently configured edge for this pin

This value will only be present if the Pin allows for interrupts.

fn get_poller(&self) -> Result<PinPoller>

Get a PinPoller object for this pin

This pin poller object will register an interrupt with the kernel and allow you to poll() on it and receive notifications that an interrupt has occured with minimal delay.