Struct sysfs_gpio::PinPoller

pub struct PinPoller { /* fields omitted */ }

Methods

impl PinPoller

pub fn get_pin(&self) -> Pin

Get the pin associated with this PinPoller

Note that this will be a new Pin object with the proper pin number.

pub fn new(pin_num: u64) -> Result<PinPoller>

Create a new PinPoller for the provided pin number

pub fn poll(&mut self, timeout_ms: isize) -> Result<Option<u8>>

Block until an interrupt occurs

This call will block until an interrupt occurs. The types of interrupts which may result in this call returning may be configured by calling set_edge() prior to making this call. This call makes use of epoll under the covers. To poll on multiple GPIOs or other event sources, poll asynchronously using the integration with either mio or tokio_core.

This function will return Some(value) of the pin if a change is detected or None if a timeout occurs. Note that the value provided is the value of the pin as soon as we get to handling the interrupt in userspace. Each time this function returns with a value, a change has occurred, but you could end up reading the same value multiple times as the value has changed back between when the interrupt occurred and the current time.